+++ src/sksundae/_cy_common.pyx
@@ -15,7 +15,13 @@ from .c_sunmatrix cimport *  # Access to SUNMatrix functions
 # the pxi file and used here.
 include "py_config.pxi"
 
-sundials_version = SUNDIALS_VERSION
+config = {
+    'SUNDIALS_VERSION': SUNDIALS_VERSION,
+    'SUNDIALS_FLOAT_TYPE': SUNDIALS_FLOAT_TYPE,
+    'SUNDIALS_INT_TYPE': SUNDIALS_INT_TYPE,
+    'SUNDIALS_SUPERLUMT_ENABLED': SUNDIALS_SUPERLUMT_ENABLED,
+    'SUNDIALS_SUPERLUMT_THREAD_TYPE': SUNDIALS_SUPERLUMT_THREAD_TYPE,
+}
 
 if SUNDIALS_FLOAT_TYPE == "float":
     from numpy import float32 as DTYPE
@@ -62,8 +68,7 @@ cdef np2ptr(np.ndarray[DTYPE_t, ndim=1] np_array, sunrealtype* nv_ptr):
 
 cdef np2smat_dense(np.ndarray[DTYPE_t, ndim=2] np_A, SUNMatrix smat):
     """Fill a SUNDenseMatrix with values from a 2D numpy array."""
-    cdef sunindextype i
-    cdef sunindextype j 
+    cdef sunindextype i, j
     cdef sunindextype M = <sunindextype> np_A.shape[0]
     cdef sunindextype N = <sunindextype> np_A.shape[1]
     cdef sunrealtype** sm_cols = SUNDenseMatrix_Cols(smat)
@@ -75,8 +80,7 @@ cdef np2smat_dense(np.ndarray[DTYPE_t, ndim=2] np_A, SUNMatrix smat):
 
 cdef np2smat_band(np.ndarray[DTYPE_t, ndim=2] np_A, SUNMatrix smat):
     """Fill a SUNBandMatrix with values from a 2D numpy array."""
-    cdef sunindextype i
-    cdef sunindextype j 
+    cdef sunindextype i, j
     cdef sunindextype N = <sunindextype> np_A.shape[1]
     cdef sunrealtype** sm_cols = SUNBandMatrix_Cols(smat)
     cdef sunindextype lband = SUNBandMatrix_LowerBandwidth(smat)
@@ -92,7 +96,52 @@ cdef np2smat_band(np.ndarray[DTYPE_t, ndim=2] np_A, SUNMatrix smat):
             sm_cols[j][i-j+smu] = np_A[i,j]
 
 
-cdef np2smat(np.ndarray[DTYPE_t, ndim=2] np_A, SUNMatrix smat):
+cdef np2smat_sparse1D(np.ndarray[DTYPE_t, ndim=1] np_A, SUNMatrix smat,
+                      object sparsity):
+    """Fill a SUNSparseMatrix with values from a 1D numpy array."""
+    cdef sunindextype nnz, nidx, nptr
+    cdef sunrealtype* data = SUNSparseMatrix_Data(smat)
+    cdef sunindextype* indices = SUNSparseMatrix_IndexValues(smat)
+    cdef sunindextype* indptrs = SUNSparseMatrix_IndexPointers(smat)
+    cdef np.ndarray[INT_TYPE_t, ndim=1] np_indices = sparsity.indices
+    cdef np.ndarray[INT_TYPE_t, ndim=1] np_indptr = sparsity.indptr
+
+    nnz = <sunindextype> sparsity.nnz
+    nidx = <sunindextype> sparsity.indices.size
+    nptr = <sunindextype> sparsity.indptr.size
+
+    data[0:nnz] = &np_A[0]
+    indices[0:nidx] = <sunindextype*> &np_indices[0]
+    indptrs[0:nptr] = <sunindextype*> &np_indptr[0]
+
+
+cdef np2smat_sparse2D(np.ndarray[DTYPE_t, ndim=2] np_A, SUNMatrix smat,
+                      object sparsity):
+    """Fill a SUNSparseMatrix with values from a 2D numpy array."""
+    cdef sunindextype i, j, start, end, idx
+    cdef sunindextype nnz, nidx, nptr, ncols
+    cdef sunrealtype* data = SUNSparseMatrix_Data(smat)
+    cdef sunindextype* indices = SUNSparseMatrix_IndexValues(smat)
+    cdef sunindextype* indptrs = SUNSparseMatrix_IndexPointers(smat)
+    cdef np.ndarray[INT_TYPE_t, ndim=1] np_indices = sparsity.indices
+    cdef np.ndarray[INT_TYPE_t, ndim=1] np_indptr = sparsity.indptr
+
+    nidx = <sunindextype> sparsity.indices.size
+    nptr = <sunindextype> sparsity.indptr.size
+
+    indices[0:nidx] = <sunindextype*> &np_indices[0]
+    indptrs[0:nptr] = <sunindextype*> &np_indptr[0]
+
+    ncols = <sunindextype> np_A.shape[1]
+    for j in range(ncols):
+        start = sparsity.indptr[j]
+        end = sparsity.indptr[j+1]
+
+        for i, idx in enumerate(sparsity.indices[start:end]):
+            data[start + i] = np_A[idx,j]
+
+
+cdef np2smat(np.ndarray np_A, SUNMatrix smat, object sparsity):
     """Fill a SUNMatrix with values from np_A using the correct cdef."""
     cdef SUNMatrix_ID matrix_id = SUNMatGetID(smat)
 
@@ -100,5 +149,10 @@ cdef np2smat(np.ndarray[DTYPE_t, ndim=2] np_A, SUNMatrix smat):
         np2smat_dense(np_A, smat)
     elif matrix_id == SUNMATRIX_BAND:
         np2smat_band(np_A, smat)
+    elif matrix_id == SUNMATRIX_SPARSE and np_A.ndim == 1:
+        np2smat_sparse1D(np_A, smat, sparsity)
+    elif matrix_id == SUNMATRIX_SPARSE and np_A.ndim == 2:
+        np2smat_sparse2D(np_A, smat, sparsity)
     else:
-        raise TypeError("'smat' must be a 'dense' or 'band' SUNMatrix.")
+        raise TypeError("Only 'dense', 'band', or 'sparse' SUNMatrix are"
+                        " supported for 'smat'.")